Fix: Indent multiline fixes (fixes #120) (#124)
* Chore: Add failing test case for indent of multiline fixes

If the replacement text of an autofix contains newlines, all lines but
the first in the replacement text will unexpectedly have 0 indentation
if a code block is indented. I guess not only the _range_ of fixes but
also the _text_ of fixes need to be adjusted.

If the code block is placed within a blockquote, a `>` character also
needs to be inserted.

This issue was found while investigating
lydell/eslint-plugin-simple-import-sort#22.  That plugin several lines
of imports in one go. This means that most of the imports will have too
little indentation after autofix if the code block is indented (such as
when placed in a list).

* Fix: Indent multiline fixes (fixes #120)

* Add test for underindented multiline autofix

* Add test for nested blockquote multiline autofix

* Add comment explaining underintented line shift