Docs: Reference recommended config disabled rules (#159)
* Docs: Relocate recommended rule section in docs

This is a cut-and-paste commit. I'm going to update these to reflect the
recommended config and want to put them under the "Advanced
Configuration" section. I did this in a precursor commit to make
reviewing the actual textual diff easier in the next commit.

* Docs: Reference recommended config disabled rules

* Docs: Explain rules disabled in recommended config

These borrow a snippet of text from the corresponding documentation in
the readme.

* Docs: Simplify wording

Thanks to @ilyavolodin for the suggestions to start this off.