Build: Install example test dependencies in Jenkins (#160)
The 2.0.0-rc.0 release attempt failed when trying to run the examples'
tests because their dependencies were never installed. In #152, I added
an install step to the GitHub Actions workflow but not the Jenkins
release flow. This change adds an `install-examples` npm script and
calls it both from the GitHub Actions workflow and the
`generate-.*release` scripts.