Chore: Switch to main for primary branch (fixes #161) (#165)
* Chore: Switch to main for primary branch (fixes #161)

The only remaining string search results for "master" are URLs to other
repositories.

* Chore: Use latest eslint-release

This should allow releasing from the `main` branch, which was part of
the 3.1 release. The breaking changes in v2 and v3 dropped support for
older Node versions on the release server. Since it's working in other
eslint org packages, that's not a concern.