Fix: npm prepare script on Windows (refs #166)
The previous script worked on macOS and Linux, but I neglected to
consider Windows. This wouldn't affect users thankfully, only
developers, so the blast radius would have been small had it not been
caught. Now that we're testing on Windows, future bugs like this should
be caught. I'm out of ideas for doing this entirely inside npm scripts,
so I broke down and did it in JS.