Fix: Use internally-calculated code block trim lengths
Future versions of the Markdown processor don't include per-line indent
information. Thankfully, we're already calculating it in
`getBlockRangeMap()` because we need it to correctly map fix ranges for
lines that are under-indented to the left of their opening code fence.
This includes the indent in the `RangeMap` for each line and uses that
in `adjustBlock()` instead.

This uncovered an incorrect assertion in one of the tests, which
combined an indented code block, a message on a line with mixed spaces
and tabs indentation, and an under-indented final line. The message
column was incorrectly asserted as being in the whitespace indentation
before the line.