ci: Work around npm behavior changes to fix CI on main (#206)
* ci: Prevent recursive prepare script invocations

A change between npm 8.9 and 8.10 started running the main package's
`prepare` script as part of the examples' installs. This broke CI when
the install step launched a fork bomb and got killed after several
minutes.

This change should fix builds on `main`.

* ci: Link with legacy peer deps

Something about peer dependencies changed between npm 8.5.5 and 8.6.0
that is causing CI to fail on `main` with ESLint 6 on the test matrix.

This works around the issue.